Update curve-curvature-new.tex
Until exercise 3.2, we have defined curvature only for unit speed curves, but it is obviously true for any curve. 
Proof of 3.6 got shortened into refrences.
Introduced before Frenchel theorem the definition for total curvature of closed curve. It is a concept that confused a lot the MASS students. Does the last angle count?
Deleted the comments after piecewise smooth curve definition, since the drawing is more than clear and the discussion is easily assumed.
Drawing of proposition 3.14 has the beta angles wrong. Still need to fix it.
When comparing two similar things like k1 and k2 are the curvatures of gamma1 and gamma 2, respectively is the correct word, instead of correspondingly.