Update plane-curves.tex
We can say 5.2 is a particular case of the space curve version with tor=0. (with a little modification to allow k<0). 
Then that we want to give an alternate proof for the existence to get corollary 5.3. I liked more the previous version of corollary 5.3 saying theta ' = k instead of the integral form.
Line 173 seemed unnecesary. 
The definition of a curve supporting another one is very complicated. On the other hand, a supporting circline is really easy to define since circlines have well defined sides. I think it was a better idea to introduce supporting circlines before defining a curve supporting another.